add new commandline parameter `--jni-generate-main` to resolve the currently discussed issue on how to provide a default JNI_OnLoad & JNI_OnUnload implementation without the user having to know about the internals.
The paramater, if true, generates a file `djinni_jni_main.cpp` that includes the `djinni/jni/djinni_jni_main.hpp` header provided by djinni-support-lib. That way by default the user automatically gets a default JNI_OnLoad implementation if he builds a shared library from all sources.

This is a breaking change because it generates a new type of file. This requires the upcoming djinni-support-lib v1.0.0, that comes with the required header file, to work properly.